refactor: remove common and move subfolders up one level
