+++ src/renderer/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Stack";
+export * from "./Button";
+export * from "./TextInput";
