+++ src/renderer/src/hooks/useCurrentTime.tsx
